[skip ci]fix typo and relocate files
